Address PR review feedback for org saved opportunity notification
